ref_soft: fix pointer truncation in software renderer

+++ ref_soft/r_misc.c
@@ -56,10 +56,10 @@ void D_Patch (void)
 
 	if (!protectset8)
 	{
-		Sys_MakeCodeWriteable ((int)D_PolysetAff8Start,
-						     (int)D_Aff8Patch - (int)D_PolysetAff8Start);
-		Sys_MakeCodeWriteable ((long)R_Surf8Start,
-						 (long)R_Surf8End - (long)R_Surf8Start);
+		Sys_MakeCodeWriteable ((uintptr_t)D_PolysetAff8Start,
+						     (uintptr_t)D_Aff8Patch - (uintptr_t)D_PolysetAff8Start);
+		Sys_MakeCodeWriteable ((uintptr_t)R_Surf8Start,
+						 (uintptr_t)R_Surf8End - (uintptr_t)R_Surf8Start);
 		protectset8 = true;
 	}
 	colormap = vid.colormap;
